refactor(next): reorganize template folder structure for clarity
Categorize the Next.js starter so the layout is easy to scan for
developers scaffolding from it:

- components/welcome/: split 14 flat files into panels/ (section
  components) and ui/ (presentational primitives); keep
  data/theme/networks/shared/format/useDismissible at the folder root
- lib/web3/: group all wallet + chain wiring in one place
  (provider, restrict-polkadot-chains, chains)
- config/: move build-time bundler stubs out of lib/
- add the standard @/* path alias (tsconfig) so imports read cleanly
  after the move
- remove dead create-next-app SVGs; keep public/ via .gitkeep
- README: add a project-structure map and refresh all paths

components/App.tsx stays put as the documented "START HERE" entry.
File moves are git renames, so history is preserved. Verified with
tsc --noEmit, eslint, and next build (all pass).
