+++ templates/next/app/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import App from "../components/App";
+import App from "@/components/App";
 
 export default function Home() {
   return <App />;
