+++ templates/next/components/welcome/shared.ts
