+++ templates/next/components/welcome/ui/icons.tsx
