+++ templates/next/config/empty-module.js
