+++ templates/next/config/metamask-analytics-stub.js
